Added Version 3.0.7
11/30/18
- added the buttons to activate the biofeedback and set bias and target

12/04/18
- Now biofeedback works. It provides on A17 the tune required depending on the frequency decided by the control.

The biofeedback
It reuires a special cable for the pot in particular we use the yellow blue red for the pot and the brow green for the jack.

12/11/18
the arduino code now has a counter for the torque error in order to be sure that the stop is launched only in case of a real disconnection
of the cable and not just by the noise on the signal

the code is now commented and two old controls never used have been deleted. If you wanna see them go to A_exo_3_0_6-To_Test
the matlab gui has not more these buttons

the relative commands that now are available are 's' 'S' 'p' 'P'. Better to double check in receive and transmit.