Added Version 3.1.0
Reinstated optimization functionality in MATLAB and Arduino codes.

Added functionality to run one and two parameter optimization in HLO code. If one parameter is sent to MATLAB and proportional
control is OFF, the bang-bang sigmoid function is used. If two parameters are sent to MATLAB and proportional control is OFF, the bang-bang zero-jerk spline is used.
If one parameter is sent to MATLAB and proportional control is ON, proportional torque is set as normal.

Changed HLO code to generate folders for each participant and date.